Override yank handlers when pasting
The pasted text may specify a yank handler, e.g., evil-yank-line-handler,
that inserts the text outside of the current overlay.  It is necessary to
override any such yank handler in order to ensure that the text replaces
the current overlay and gets propagated to the other overlays.

* evil-iedit-state.el (evil-iedit-state/paste-replace): Specify
insert-for-yank as the yank handler when calling evil-paste-before.