iedit-cleanup -> iedit-lib-cleanup?
Hi, how are things?

Call to iedit-cleanup in iedit-done (line 194) seems to cause "Symbol's function definition is void: iedit-cleanup" in 0.300.0@26.3 (spacemacs) when trying to exit iedit-mode (ESC-ESC or C-g). The ".emacs.d" is under "develop".

If iedit-cleanup is replaced with iedit-lib-cleanup and iedit-done re-evaluated, the issue is gone.

Hope this is useful. Thank you!

/VT